requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
docker==6.0.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
packaging==23.0
py-cord==2.4.0
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
websocket-client==1.5.1
yarl==1.8.2
aiofiles==0.3.2
aiosqlite==0.17.0
iso8601==1.1.0
pypika-tortoise==0.1.6
pytz==2022.7.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
